project-assignment2/requirements.txt
requests
certifi
confluent-kafka[avro,json,protobuf]>=1.4.2
psycopg2-binary